requirements.txt
bitsandbytes==0.44.1
numpy==2.1.3
pandas==2.2.3
pyarrow==18.1.0
pyresidfp==0.14.0
pytorch-lightning==2.4.0
scipy==1.14.1
schedulefree==1.3
tensorboard==2.18.0
torchtune==0.4.0
torchao==0.6.1
